Lose the custom SSH port
This was a remnant of an old configuration where I didn't use a VPN. At
the time, it made sense to move sshd to a different port to deter dumb
bots trying to bruteforce their way in. Since sshd is long since hidden
behind a VPN, this setting is obsolete and only complicates the ansible
scripts.