Revert grafana to listen on localhost
Making grafana listen on a Unix socket would require to set the socket's
gid to that of a group the caddy user is a member of. For grafana to be
able to do that, the grafana user must also be a member of that group. I
don't want to add grafana to the caddy group for security reasons. The
remaining option would be to define a new group with grafana and caddy
as members, but then I run into an idiotic Nix issue where unspecified
GIDs are created at *activation* time, not at *build* time. That way,
the GID is not available to be passed to grafana, and everything breaks.
I _could_ specify a static GID, but then I run into risks of conflicts.